Bugfix where source demo would end too soon and switched from median ping to lowest ping (since that gives a more realistic estimation of delay between real client and server and less interference of packet received/snap sent delay)

+++ shared/DemoReader.cpp
@@ -532,7 +532,7 @@ int DemoReader::getCurrentDemoTime() {
 	return demoCurrentTime;
 }
 
-qboolean DemoReader::SeekToTime(int time) {
+qboolean DemoReader::SeekToTime(float time) {
 	while (demoCurrentTime < time && !endReached) {
 		ReadMessage();
 	}
